Backend/requirements.txt
fastapi
uvicorn
pydantic
openai
python-multipart
PyJWT
python-jose[cryptography]
passlib[bcrypt]